requirements.txt
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
backports.functools-lru-cache==1.6.5
beautifulsoup4==4.12.2
bleach==6.1.0
boto==2.49.0
Brotli==1.0.9
cached-property==1.5.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.2
comm==0.1.4
confluent-kafka==2.3.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distro==1.8.0
entrypoints==0.4
exceptiongroup==1.1.3
executing==2.0.1
fastjsonschema==2.18.1
fqdn==1.5.1
frozendict==2.3.9
h11==0.14.0
html5lib==1.1
httpcore==1.0.1
httpx==0.25.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipykernel==6.26.0
ipython==8.17.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.0
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.1
lxml==4.9.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==3.0.2
multitasking==0.0.11
nbclient==0.8.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
notebook_shim==0.2.3
numpy==1.26.1
oauthlib==3.2.2
openai==1.1.2
overrides==7.4.0
packaging==23.2
pandas==2.1.2
pandocfilters==1.5.0
parso==0.8.3
peewee==3.17.0
pexpect==4.8.0
pickleshare==0.7.5
pip==23.3
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
praw==7.7.1
prawcore==2.4.0
prometheus-client==0.18.0
prompt-toolkit==3.0.39
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
PySocks==1.7.1
pyspark==3.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.0
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
Send2Trash==1.8.2
setuptools==68.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.2
terminado==0.17.1
tiktoken==0.5.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
tweepy==4.14.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
typing-utils==0.1.0
tzdata==2023.3
update-checker==0.18.0
uri-template==1.3.0
urllib3==2.0.7
wcwidth==0.2.9
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
wheel==0.41.2
yfinance==0.2.32
zipp==3.17.0
